Create README.md
this is a webapp for speading awarness for different disease when outbreak occurs. A database prepared by government is used and i particular state disease bihar is taken here with different districts of bihar for eg. patna had a outbreak of chicken pox in 2019. Apis used is google Map map tool. whenever a person clicks on that particular place on map he will be directed to a prevetion page which includes measures of prevention and vedios of that particular disease. As we all know "prevention is beeter than cure" . For making it real time we have used report from doctors where a registered and verified doctor can enter no cases of a disease in a particular day . that data will go in excel sheet which can be pushed into maps and plotted there .Its a multi lingual app